do not duplicate list of tools, make it clear this is optional
We should avoid documentation that duplicates code where possible,
and in this instance the code had better comments than the docs.
By pointing people to the source of truth, we avoid the need to
make two updates whenever the list changes.

Also trying to signpost that folks don't _need_ to install all the
things, as it's not very YAGNI/JIT. Tools will be naturally
selected by their usefulness and through pairing. It does
however make sense to provide the list for convenience and for the
comfort of those who prefer to be prepared.